requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.0
crispy-bootstrap5==0.7
cryptography==41.0.4
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==4.2.6
django-braces==1.15.0
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-countries==7.5.1
django-crispy-forms==2.1
django-jquery==3.1.0
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.1.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
fontawesomefree==6.4.2
idna==3.4
jwcrypto==1.5.0
oauthlib==3.2.2
phonenumbers==8.13.35
Pillow==10.1.0
psycopg==3.1.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==1.7.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wrapt==1.15.0
